+++ R/Distribution.R
@@ -430,11 +430,11 @@ Distribution <- R6Class("Distribution", lock_objects = FALSE,
 
       data <- pdq_point_assert(..., self = self, data = data)
       if (inherits(data, "matrix")) {
-        assert(self$liesInType(apply(data, 1, as.Tuple), all = TRUE, bound = FALSE),
+        assert(self$liesInType(apply(data, 1, as.Tuple), all = TRUE, bound = TRUE),
                .var.name = "Do all points lie in Distribution domain?"
         )
       } else {
-        assert(self$liesInType(as.numeric(data), all = TRUE, bound = FALSE),
+        assert(self$liesInType(as.numeric(data), all = TRUE, bound = TRUE),
                .var.name = "Do all points lie in Distribution domain?"
         )
       }
@@ -489,11 +489,11 @@ Distribution <- R6Class("Distribution", lock_objects = FALSE,
 
       data <- pdq_point_assert(..., self = self, data = data)
       if (inherits(data, "matrix")) {
-        assert(self$liesInType(apply(data, 1, as.Tuple), all = TRUE, bound = FALSE),
+        assert(self$liesInType(apply(data, 1, as.Tuple), all = TRUE, bound = TRUE),
                .var.name = "Do all points lie in Distribution domain?"
         )
       } else {
-        assert(self$liesInType(as.numeric(data), all = TRUE, bound = FALSE),
+        assert(self$liesInType(as.numeric(data), all = TRUE, bound = TRUE),
                .var.name = "Do all points lie in Distribution domain?"
         )
       }